requirements.txt
aiohttp==3.7.4.post0
astroid==2.5.6
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
chardet==4.0.0
discord.py==1.7.2
flake8==3.9.2
idna==3.1
isort==5.8.0
jedi==0.17.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
mypy==0.812
mypy-extensions==0.4.3
mysql-connector-python==8.0.25
parso==0.7.1
pluggy==0.13.1
protobuf==3.17.0
pycodestyle==2.6.0
pydocstyle==6.1.1
pyflakes==2.2.0
pylint==2.8.2
pyls-mypy==0.1.8+6.g3b105b4
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
rope==0.19.0
six==1.16.0
snowballstemmer==2.1.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
ujson==4.0.2
wrapt==1.12.1
yapf==0.31.0
yarl==1.6.3
